# rates-api-spec.yml
openapi: 3.0.3
info:
  title: Rates Api
  description: Handle basic rates operations
  version: "1.0"
servers:
  - url: http://localhost:8080/
    description: Sandbox server to test basic rates operations
tags:
- name: rates
  description: Available operations for rates domain.
paths:
  /rates/{id}:
    get:
      tags:
      - rates
      summary: Get rate detail
      operationId: getRateById
      parameters:
        - in: path
          description: The rate identifier
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rate'
        404:
          description: Rate not found
    put:
      tags:
      - rates
      summary: Update a rate or create if doesn't exists
      operationId: upsertRate
      parameters:
        - in: path
          name: id
          description: The rate identifier
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Required info for a upsert Rate
        content:
          application/json:
            schema:
              type: object
              properties:
                brandId:
                  type: string
                  format: uuid
                productId:
                  type: string
                  format: uuid
                startDate: 
                  type: string
                  format: date
                  example: "2021-01-10"
                endDate:
                  type: string
                  format: date
                  example: "2021-01-10"
                price:
                  type: integer
                currencyCode:
                  type: string
                  maxLength: 3
                  minLength: 3
      responses:
        201:
          description: rate created successfully
        204:
          description: rate updated successfully
        default:
          description: unexpected error occured
      
components:
  schemas:
    Rate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        brandId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
          description: Rate initial date
          example: "2021-01-01"
        endDate:
          type: string
          format: date
          description: Rate ending date
          example: "2021-03-10"        
        price:
          type: number
          format: double
          example: 1050.23
        currencySymbol:
          type: string
          minLength: 1
          maxLength: 1
        currencyCode:
          type: string
          minLength: 3
          maxLength: 3
      required:
        - id
        - brandId
        - productId
        - startDate
        - endDate
        - price
        - currencySymbol
        - currencyCode